Increased test timeout + clean test working dir before testing
Recently tests (particularly the windows ones) have been failing because they would not finish fast enough for a timeout. This increases the timeout so that they have time to finish.

Also cleaned the `test/work` directory before running tests so that tests that are force-stopped mid-test do not screw up future tests. Added `test/work` to the standard ignore list as well.