Remove hard dependancy on electron-prebuilt
This means packager will also use electron version specified in `electron-version` field in package.json

This is good because we don't have to depend on `electron-prebuilt`. `prebuilt` automatically downloads
electron for the current platform, but `packager` might only be building for a different platform, and will use
`electron-download` anyway, so we can avoid unnecessary download, and speed up `npm install` time.