Add ARM64 support (#711)
With https://github.com/electron/electron/pull/10219 merged, ARM64
support is likely to be official starting with the Electron 1.8 release
series. This adds basic official support for the arch, plus adds it to
the platform/arch combinations when `all` is true or `platform=linux`
and `arch=all` is set, and the specified version of Electron is 1.8.0
or above.