Moved hostArch() earlier in validateListFromOptions()
This PR is in response to atom/atom#15431.  When I tried to build Atom
on my Chromebook running Arch Linux ARM, the automatic trigger detecting
armv7l was not running because a value already existed for the arch.

This change moves the hostArch() function before the list assignment to
catch this case.  When I run this code on my Chromebook, the packaging
code appears to succeed.